Move Diagnose to Typology
